requirements.txt
phidata
python-dotnev
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
streamlit